LayoutTests/imported/w3c/web-platform-tests/css/css-align/blocks/align-content-table-cell-expected.txt

first
second
first
second
first
second
third
first
second
first
second
first
second
third

PASS vertical-align:top and align-content:start are equivalent
FAIL vertical-align:middle and `align-content:unsafe center` are equivalent assert_equals: expected 10 but got 11
FAIL vertical-align:bottom and `align-content:unsafe end` are equivalent assert_equals: expected 30 but got 29
PASS vertical-align:baseline and align-content:baseline are equivalent
PASS vertical-align:middle and `align-content:safe center` are equivalent if the container is tall
FAIL vertical-align:bottom and `align-content:safe end` are equivalent if the container is tall assert_equals: expected 30 but got 29
PASS `align-content:safe center` is equivalent to `unsafe center` even if the specified `height` is short
PASS `align-content:safe end` is equivalent to `unsafe end` even if the specified `height` is short
